+++ backend/thebookclub/src/main/java/com/projectg2/thebookclub/model/TipoPagamento.java
@@ -1,4 +1,4 @@
-package br.com.digitalhouse.thebookclub.modelo;
+package com.projectg2.thebookclub.model;
 
 public enum TipoPagamento {
